Feat/29 media menu n group chat

+++ src/components/new/Navigate.tsx
@@ -1,5 +1,6 @@
+/* eslint-disable @typescript-eslint/no-misused-promises */
 import clsx from 'clsx';
-import React from 'react';
+import { lazy, MouseEvent, useEffect, useRef, useState } from 'react';
 import { BsPerson } from 'react-icons/bs';
 import { IoMdRemoveCircleOutline } from 'react-icons/io';
 import { IoChatbubbleOutline } from 'react-icons/io5';
@@ -8,37 +9,74 @@ import { PiDotsNineBold, PiGearSixBold } from 'react-icons/pi';
 import { useNavigate } from 'react-router-dom';
 
 import { useAppDispatch } from '../../hooks';
+import { useConfirm } from '../../hooks/useConfirm';
 import { useDeleteUser } from '../../hooks/useDeleteUser';
 import { useLogout } from '../../hooks/useLogout';
-import { setSetting } from '../../store';
+import { useSetTheme } from '../../hooks/useSetTheme';
+import { socket } from '../../service/socket';
+import { clearCurrentConversation, setSetting } from '../../store';
+import { clearNewConversation } from '../../store/new-conversation-slice';
 import Icon from '../atoms/Icon';
 import Spinner from '../atoms/Spinner';
 import Contacts from './nav/Contacts';
 import Conversations from './nav/Conversations';
 import SearchBox from './nav/SearchBox';
 
+const User = lazy(() => import('../User'));
+
+type UpdateConversationType = [
+  {
+    conversationId: string;
+    name: string;
+    creator: string | null;
+    isGroup: boolean;
+    avatar: string | null;
+    createdAt: string;
+    lastMessage: string;
+    lastMessageAt: string;
+    isLastMessageSeen: boolean;
+    status: 'online' | 'offline';
+    totalUnreadMessages: number;
+    participants: {
+      id: string;
+      avatar: string;
+    }[];
+  },
+];
+
 export default function Navigate() {
-  const [shouldSettingOpen, setSettingOpen] = React.useState<boolean>(false);
-  const buttonSettingRef = React.useRef<HTMLButtonElement | null>(null);
-  const settingMenuRef = React.useRef<HTMLDivElement | null>(null);
+  const [shouldSettingOpen, setSettingOpen] = useState<boolean>(false);
+  const buttonSettingRef = useRef<HTMLButtonElement | null>(null);
+  const settingMenuRef = useRef<HTMLDivElement | null>(null);
   const navigate = useNavigate();
+  const confirm = useConfirm();
   const { mutate, isPending } = useLogout();
+  const { mutate: setTheme } = useSetTheme();
   const { mutate: deleteUser, isPending: isPendingDeleteUser } = useDeleteUser();
-  const handleDeleteUser = (event: React.MouseEvent<HTMLButtonElement, UIEvent>) => {
+  const handleDeleteUser = async (event: MouseEvent<HTMLButtonElement, globalThis.UIEvent>) => {
     event.preventDefault();
-    deleteUser();
+    const choice = await confirm({
+      isOpen: true,
+      buttonLabel: 'Delete',
+      description: 'Are you sure want to delete your account?',
+    });
+    if (choice) {
+      deleteUser();
+    }
   };
   const dispatch = useAppDispatch();
-  const handleLogout = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+  const handleLogout = (event: MouseEvent<HTMLButtonElement, globalThis.MouseEvent>) => {
     event.preventDefault();
     mutate();
   };
-  const handleNewConversation = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+  const handleNewConversation = (event: MouseEvent<HTMLButtonElement, globalThis.MouseEvent>) => {
     event.preventDefault();
+    dispatch(clearNewConversation());
+    dispatch(clearCurrentConversation());
     navigate('new');
   };
-  React.useEffect(() => {
-    const handler = (event: MouseEvent) => {
+  useEffect(() => {
+    const handler = (event: globalThis.MouseEvent) => {
       if (
         !(
           settingMenuRef.current?.contains(event.target as HTMLElement) ||
@@ -53,13 +91,20 @@ export default function Navigate() {
       document.removeEventListener('click', handler);
     };
   }, []);
+  useEffect(() => {
+    socket.on('update conversations', (args: UpdateConversationType) => console.log(args));
+  }, []);
   return (
-    <div className="w-[25%] h-full px-2 py-8 gap-6 flex flex-col bg-[#221f34]">
+    <div className="w-[25%] h-full px-2 py-8 gap-6 flex flex-col bg-surface-mix-200">
+      <User />
       <div className="flex gap-2 w-full pr-4 items-center justify-center">
         <SearchBox />
-        <div className="w-[10%] h-full">
-          <button className="btn bg-[#343142] hover:bg-[#343142] m-0 " onClick={handleNewConversation}>
-            <Icon className="text-2xl">
+        <div className="w-[15%] h-full">
+          <button
+            className="w-full h-full rounded-lg flex items-center justify-center bg-surface-mix-100 hover:bg-surface-mix-300 outline-none border-none m-0 "
+            onClick={handleNewConversation}
+          >
+            <Icon className="text-3xl">
               <MdOpenInNew />
             </Icon>
           </button>
@@ -76,7 +121,7 @@ export default function Navigate() {
       <div className="w-full flex gap-2 relative ">
         <div className="w-full ">
           <button
-            className="btn btn-error btn-sm w-full text-white hover:bg-red-500 hover:outline-none hover:border-none disabled:!bg-red-500 disabled:!cursor-not-allowed disabled:text-white"
+            className="btn btn-error btn-sm w-full text-color-base-100 hover:bg-red-500 hover:outline-none hover:border-none disabled:!bg-red-500 disabled:!cursor-not-allowed disabled:text-color-base-100"
             onClick={handleLogout}
             disabled={isPending}
           >
@@ -97,13 +142,13 @@ export default function Navigate() {
         <div
           ref={settingMenuRef}
           className={clsx(
-            'absolute  bottom-10  z-10 right-0 p-2 inline-block text-sm font-medium bg-gray-600/40 border-none rounded-xl dark:text transition-all  duration-900 ease-in-out  w-44 h-auto origin-bottom-right',
+            'absolute  bottom-10  z-10 right-0 p-2 inline-block text-sm font-medium bg-surface-mix-300 border-none rounded-xl dark:text transition-transform  duration-900 ease-in-out  w-44 h-auto origin-bottom-right',
             !shouldSettingOpen ? ' opacity-0 scale-0' : 'opacity-100 scale-100  ',
           )}
         >
           <button
             type="button"
-            className="w-full px-2 py-2 font-medium text-left rounded-[8px] border-gray-200 cursor-pointer hover:bg-red-700 text-white focus:outline-none flex items-center gap-2 disabled:"
+            className="w-full px-2 py-2 font-medium text-left rounded-[8px] border-gray-200 cursor-pointer hover:bg-red-600 hover:text-white transition-colors hover:text-color-base-100 text-color-base-100 focus:outline-none flex items-center gap-2 disabled:"
             onClick={handleDeleteUser}
             disabled={isPendingDeleteUser}
           >
@@ -122,7 +167,20 @@ export default function Navigate() {
           </button>
           <button
             type="button"
-            className="w-full px-2 py-2 font-medium text-left rounded-[8px] border-gray-200 cursor-pointer hover:bg-gray-700 text-white focus:outline-none flex items-center gap-2"
+            className="w-full px-2 py-2 font-medium text-left rounded-[8px] border-gray-200 cursor-pointer hover:bg-surface-mix-400 text-color-base-100 focus:outline-none flex items-center gap-2"
+            onClick={() => {
+              const e = document.getElementsByTagName('body');
+              if (e.length > 0) {
+                const theme = e[0].getAttribute('data-theme');
+                if (theme === 'dark') {
+                  e[0].setAttribute('data-theme', 'light');
+                  setTheme('light');
+                } else {
+                  e[0].setAttribute('data-theme', 'dark');
+                  setTheme('dark');
+                }
+              }
+            }}
           >
             <Icon className="text-xl">
               <MdOutlineDarkMode />
@@ -131,7 +189,7 @@ export default function Navigate() {
           </button>
           <button
             type="button"
-            className="w-full px-2 py-2 font-medium text-left rounded-[8px] border-gray-200 cursor-pointer hover:bg-gray-700 text-white focus:outline-none flex items-center gap-2"
+            className="w-full px-2 py-2 font-medium text-left rounded-[8px] border-gray-200 cursor-pointer hover:bg-surface-mix-400 text-color-base-100 focus:outline-none flex items-center gap-2"
             onClick={(e) => {
               e.preventDefault();
               dispatch(setSetting('general'));
@@ -145,7 +203,7 @@ export default function Navigate() {
           </button>
           <button
             type="button"
-            className="w-full px-2 py-2 font-medium text-left rounded-[8px] border-gray-200 cursor-pointer hover:bg-gray-700 text-white focus:outline-none flex items-center gap-2"
+            className="w-full px-2 py-2 font-medium text-left rounded-[8px] border-gray-200 cursor-pointer hover:bg-surface-mix-400 text-color-base-100 focus:outline-none flex items-center gap-2"
             onClick={(e) => {
               e.preventDefault();
               dispatch(setSetting('account'));
